+++ examples/wstrust/test/src/test/java/org/jboss/jaxws/ServiceWstrustOpenShiftJarApplication.java
@@ -25,7 +25,7 @@ import org.apache.maven.settings.building.SettingsBuildingException;
 import org.eclipse.aether.resolution.ArtifactResolutionException;
 import org.jboss.intersmash.application.openshift.BootableJarOpenShiftApplication;
 import org.jboss.intersmash.application.openshift.input.BinarySource;
-import org.jboss.intersmash.test.deployments.util.maven.ArtifactProvider;
+import org.jboss.intersmash.util.maven.ArtifactProvider;
 
 import cz.xtf.builder.builders.SecretBuilder;
 import cz.xtf.builder.builders.secret.SecretType;
